+++ src/core/tasksRegistry.js
@@ -1,6 +1,7 @@
 import { latestOverlayFeature, reprocessAllTiles } from "../features/latest-overlay/index.js";
 import { createDebouncedTask } from "./createDebouncedTask.js";
 import { threadOverlayFeature } from "../features/thread-overlay/index.js";
+import { debugLog } from "./logger.js";
 export const debouncedProcessThreadTags = createDebouncedTask(() => {
   threadOverlayFeature.sync();
 }, 100);
@@ -10,6 +11,7 @@ export const debouncedProcessAllTiles = createDebouncedTask(() => {
 }, 200);
 
 export const debouncedProcessAllTilesReset = createDebouncedTask(() => {
+  debugLog("Task Registry", "Reprocess all tiles triggered");
   reprocessAllTiles();
 }, 200);
 export const debouncedProcessAllTilesDisable = createDebouncedTask(() => {